+++ constant.js
@@ -1,0 +1,2 @@
+export const arrWeapon = ['Rock', 'Paper', 'Scissors'];
+export const winningScore = 5;
