+++ index.html
@@ -25,14 +25,14 @@
         <p class="board"></p>
       </div>
       <div class="weapon-choose">
-        <button class="rock" onclick="setPlayerSelection('Rock')">
-          <img src="rock-font.png" alt="" />
+        <button class="rock">
+          <img src="img/rock-font.png" alt="" />
         </button>
-        <button class="paper" onclick="setPlayerSelection('Paper')">
-          <img src="paper-font.png" alt="" />
+        <button class="paper">
+          <img src="img/paper-font.png" alt="" />
         </button>
-        <button class="scissors" onclick="setPlayerSelection('Scissors')">
-          <img src="scissors-font.png" alt="" />
+        <button class="scissors">
+          <img src="img/scissors-font.png" alt="" />
         </button>
       </div>
       <div class="start-game">
@@ -41,5 +41,5 @@
       </div>
     </div>
   </body>
-  <script src="script.js"></script>
+  <script src="./main.js" type="module"></script>
 </html>
