+++ style.css
@@ -31,7 +31,7 @@ body{
     left: 0;
     width: 100%;
     height: 100%;
-    background-image: url('scissors.webp');
+    background-image: url('img/scissors.webp');
     background-repeat: no-repeat;
     background-size: cover;
     opacity: 0.56;
